Simple improvements: error, RE, write encoding
With this commit I'm fixing a major problem which was preventing
the smooth execution when fetching URLS by moving the "try" to the
right place, so fetch would remain working after a redirect failure.

RE now is a variable, fixed also with r'' prefix.

UTF-8 encoding when writing the file.